Initial public release of puppet-ssh.
This is a useful helper module that I wrote. It's finally time for it to
see the light of day! It includes classes and types for easy root key
exchange / pairing. These types could be expanded with additional
functionality, but it's not needed at this time. This also includes
ssh::file support, which is very magic!