Allow ssh::recv to export duplicate sshkey's.
There is a very narrow use case for this feature, but it is in fact
necessary for the puppet-ipa multi-master replication case. It turns out
that since one host has a type to export its sshkey for each host it is
exporting to, this causes an exported resources duplicate.

Keep in mind that this doesn't let you necessarily mix true/false $fast.