Ensure the exported ssh::file::hash::wrapper has a unique $name.
The $name of the @@ssh::file::hash wrapper in ssh::file::pull didn't have
a unique name that depended on the type $name and in some cases a
collision was (apparently) possible.

It surprises me that this is possible, and if there's a day when I have
a lot of time to trace the specific details of this and see if it's
entirely necessary to need it, then I'd like to because I seem to
occasionally cause this collision in puppet-ipa, but it seem transient
when a new host is "joining" the cluster and the algorithmic topology is
shifting...

It's also possible that it would be possible to somehow work around this
problem by being more careful with the topology code in puppet-ipa and
by perhaps even introducing an FSM, but I don't think any of this is
really necessary at this time because it's overkill and this patch isn't
harmful, it's in the worst case, sub optimal.

In any case, hopefully it helps :)